Make agent tests more Galactic.
- Move unit tests out of integration test file and into test/unit.
- Move remaining "API" tests into test/integration because they use ``self.app`` and mock out the API - they are what we would call integration tests in Galaxy.
- Have unit and integration tests use the same common environment variables all now prefixed with GALAXY_TEST_ (we do this for object store stuff).
- Small tweak to orchestrator.py to make the unit tests pass.
- Clean up the mocked tests method to work cleanly without any infrastructure - some tests that used the LLM had gotten in there. Tracked the tests I removed as https://github.com/galaxyproject/galaxy/issues/21437.
