--- lib/galaxy_test/api/test_agents.py
@@ -1,1074 +1,0 @@
-"""Test Galaxy AI agents API and functionality.
-
-This module contains two test suites:
-1. Mocked tests - Deterministic tests with mocked LLM responses (always run)
-2. Live LLM tests - Integration tests requiring configured LLM (optional, marked with @pytest.mark.requires_llm)
-
-## Running the tests:
-
-### Unit tests only (no Galaxy instance needed):
-    pytest lib/galaxy_test/api/test_agents.py::TestAgentUnitMocked -v
-    pytest lib/galaxy_test/api/test_agents.py::TestAgentUnitLiveLLM -v  # requires LLM
-
-### API tests (Galaxy test instance auto-configured):
-    # Run mocked API tests (Galaxy test framework handles setup):
-    pytest lib/galaxy_test/api/test_agents.py::TestAgentsApiMocked -v
-
-    # Run live LLM API tests:
-    GALAXY_TEST_ENABLE_LIVE_LLM=1 pytest lib/galaxy_test/api/test_agents.py::TestAgentsApiLiveLLM -v
-
-### All mocked tests:
-    pytest lib/galaxy_test/api/test_agents.py -k Mocked -v
-
-### All live LLM tests:
-    GALAXY_TEST_ENABLE_LIVE_LLM=1 pytest lib/galaxy_test/api/test_agents.py -k LiveLLM -v
-
-### Configuration for live API tests (TestAgentsApiLiveLLM):
-    export GALAXY_TEST_AI_API_KEY="your-api-key"
-    export GALAXY_TEST_AI_MODEL="llama-4-scout"
-    export GALAXY_TEST_AI_API_BASE_URL="http://localhost:4000/v1/"
-    export GALAXY_TEST_ENABLE_LIVE_LLM=1
-
-### Configuration for live unit tests (TestAgentUnitLiveLLM):
-    export GALAXY_AI_API_KEY="your-api-key"
-    export GALAXY_AI_MODEL="llama-4-scout"
-    export GALAXY_AI_API_BASE_URL="http://localhost:4000/v1/"
-    export GALAXY_TEST_ENABLE_LIVE_LLM=1
-"""
-
-import logging
-import os
-from typing import (
-    Any,
-    Dict,
-)
-from unittest import mock
-from unittest.mock import (
-    AsyncMock,
-    MagicMock,
-    patch,
-)
-
-import pytest
-from pydantic_ai.models.test import TestModel
-
-from galaxy.agents import (
-    agent_registry,
-    CustomToolAgent,
-    ErrorAnalysisAgent,
-    GalaxyAgentDependencies,
-    QueryRouterAgent,
-)
-from galaxy.tool_util_models import UserToolSource
-from galaxy_test.base.populators import (
-    DatasetPopulator,
-    WorkflowPopulator,
-)
-from ._framework import ApiTestCase
-
-log = logging.getLogger(__name__)
-
-# Skip live LLM tests unless explicitly enabled
-pytestmark_live_llm = pytest.mark.skipif(
-    not os.environ.get("GALAXY_TEST_ENABLE_LIVE_LLM"),
-    reason="Live LLM tests disabled. Set GALAXY_TEST_ENABLE_LIVE_LLM=1 to enable.",
-)
-
-
-# ============================================================================
-# MOCKED TEST SUITE - Always runs, no LLM required
-# ============================================================================
-
-
-class TestAgentsApiMocked(ApiTestCase):
-    """Test the Galaxy AI agents API with mocked LLM responses.
-
-    These tests use mocked LLM responses for deterministic testing.
-    They always run in CI and don't require any LLM configuration.
-    """
-
-    def setUp(self):
-        super().setUp()
-        self.dataset_populator = DatasetPopulator(self.galaxy_interactor)
-        self.workflow_populator = WorkflowPopulator(self.galaxy_interactor)
-
-    def test_list_agents(self):
-        """Test listing available agents (no LLM needed)."""
-        response = self._get("ai/agents")
-        self._assert_status_code_is_ok(response)
-        data = response.json()
-        assert "agents" in data
-        agents = data["agents"]
-        assert len(agents) > 0
-        # Check that core agents are registered
-        agent_types = [a["agent_type"] for a in agents]
-        assert "router" in agent_types
-        assert "custom_tool" in agent_types
-        assert "error_analysis" in agent_types
-        assert "gtn_training" in agent_types
-
-    @patch("galaxy.agents.router.Agent")
-    def test_query_agent_auto_routing_mocked(self, mock_agent_class):
-        """Test automatic agent routing with mocked LLM."""
-        # Set up mock agent
-        mock_agent = AsyncMock()
-        mock_agent_class.return_value = mock_agent
-
-        # Mock routing decision
-        async def mock_run(query, *args, **kwargs):
-            result = MagicMock()
-            if "BWA" in query or "tool" in query.lower():
-                result.data = {
-                    "primary_agent": "custom_tool",
-                    "reasoning": "Tool creation request detected",
-                    "confidence": 0.95,
-                }
-            else:
-                result.data = {
-                    "primary_agent": "tool_recommendation",
-                    "reasoning": "General query",
-                    "confidence": 0.7,
-                }
-            return result
-
-        mock_agent.run = mock_run
-
-        response = self._post(
-            "ai/agents/query",
-            data={
-                "query": "Create a BWA-MEM tool for paired-end reads",
-                "agent_type": "auto",
-            },
-            json=True,
-        )
-        self._assert_status_code_is_ok(response)
-        data = response.json()
-        assert "routing_info" in data
-        assert data["routing_info"]["selected_agent"] == "custom_tool"
-
-    @patch("galaxy.agents.custom_tool.Agent")
-    def test_query_custom_tool_agent_mocked(self, mock_agent_class):
-        """Test the custom tool agent with mocked LLM."""
-        mock_agent = AsyncMock()
-        mock_agent_class.return_value = mock_agent
-
-        # Mock tool creation with UserToolSource
-        mock_tool = UserToolSource(
-            **{
-                "class": "GalaxyUserTool",
-                "id": "line-counter",
-                "name": "Line Counter",
-                "version": "1.0.0",
-                "description": "Counts lines in a file",
-                "container": "ubuntu:latest",
-                "shell_command": "wc -l < $(inputs.input_file.path) > output.txt",
-                "inputs": [],
-                "outputs": [],
-            }
-        )
-
-        async def mock_run(*args, **kwargs):
-            result = MagicMock()
-            result.data = mock_tool
-            result.output = mock_tool
-            return result
-
-        mock_agent.run = mock_run
-
-        response = self._post(
-            "ai/agents/query",
-            data={
-                "query": "Create a simple tool that counts lines in a file",
-                "agent_type": "custom_tool",
-            },
-            json=True,
-        )
-        self._assert_status_code_is_ok(response)
-        data = response.json()
-        assert "metadata" in data
-        assert data["metadata"]["tool_id"] == "line-counter"
-        assert "wc -l" in data["metadata"]["tool_yaml"]
-
-    @patch("galaxy.agents.error_analysis.Agent")
-    def test_query_error_analysis_agent_mocked(self, mock_agent_class):
-        """Test the error analysis agent with mocked LLM."""
-        mock_agent = AsyncMock()
-        mock_agent_class.return_value = mock_agent
-
-        # Mock error analysis
-        async def mock_run(*args, **kwargs):
-            from galaxy.agents.error_analysis import ErrorAnalysisResult
-
-            result = MagicMock()
-            mock_analysis = ErrorAnalysisResult(
-                error_category="tool_configuration",
-                error_severity="medium",
-                likely_cause="The command 'samtools' was not found in PATH",
-                solution_steps=[
-                    "Install samtools using conda",
-                    "Check tool dependencies",
-                    "Verify container configuration",
-                ],
-                confidence="high",
-            )
-            result.data = mock_analysis
-            result.output = mock_analysis
-            return result
-
-        mock_agent.run = mock_run
-
-        # Mock job ID
-        mock_job_id = "test_job_123"
-
-        response = self._post(
-            "ai/agents/query",
-            data={
-                "query": "Why did my job fail?",
-                "agent_type": "error_analysis",
-                "context": {"job_id": mock_job_id},
-            },
-            json=True,
-        )
-        self._assert_status_code_is_ok(response)
-        data = response.json()
-        assert "content" in data
-        # Should mention the error type or solution
-        content = data["content"].lower()
-        assert "command" in content or "samtools" in content or "not found" in content
-
-
-class TestAgentUnitMocked:
-    """Unit tests for agent implementations."""
-
-    def setup_method(self):
-        """Set up mock dependencies for each test."""
-        self.mock_config = mock.Mock()
-        self.mock_config.ai_api_key = "test-key"
-        self.mock_config.ai_model = "llama-4-scout"
-        self.mock_config.ai_api_base_url = "http://localhost:4000/v1/"
-
-        self.mock_user = mock.Mock()
-        self.mock_user.id = 1
-        self.mock_user.username = "test_user"
-
-        self.mock_trans = mock.Mock()
-        self.mock_trans.app.config = self.mock_config
-        self.mock_trans.user = self.mock_user
-
-        self.deps = GalaxyAgentDependencies(
-            trans=self.mock_trans,
-            user=self.mock_user,
-            config=self.mock_config,
-            job_manager=None,
-        )
-
-    def test_agent_config_fallback_chain(self):
-        """Test per-agent configuration with fallback logic."""
-        # Set up mock config with inference_services
-        self.mock_config.inference_services = {
-            "default": {
-                "model": "gpt-4o-mini",
-                "temperature": 0.7,
-                "max_tokens": 2000,
-            },
-            "custom_tool": {
-                "model": "claude-sonnet-4-5",
-                "temperature": 0.3,
-                "max_tokens": 3000,
-            },
-        }
-
-        # Test agent with specific config
-        custom_tool_agent = CustomToolAgent(self.deps)
-        assert custom_tool_agent._get_agent_config("model") == "claude-sonnet-4-5"
-        assert custom_tool_agent._get_agent_config("temperature") == 0.3
-        assert custom_tool_agent._get_agent_config("max_tokens") == 3000
-
-        # Test agent that falls back to default
-        error_agent = ErrorAnalysisAgent(self.deps)
-        assert error_agent._get_agent_config("model") == "gpt-4o-mini"
-        assert error_agent._get_agent_config("temperature") == 0.7
-        assert error_agent._get_agent_config("max_tokens") == 2000
-
-        # Test fallback to global config when inference_services not set
-        self.mock_config.inference_services = None
-        router_agent = QueryRouterAgent(self.deps)
-        assert router_agent._get_agent_config("model") == "llama-4-scout"  # From ai_model
-        assert router_agent._get_agent_config("api_key") == "test-key"  # From ai_api_key
-
-        # Test custom default value
-        assert router_agent._get_agent_config("temperature", 0.5) == 0.5
-        assert router_agent._get_agent_config("max_tokens", 1500) == 1500
-
-    @pytest.mark.asyncio
-    async def test_router_agent_routing_decisions(self):
-        """Test that router agent makes correct routing decisions."""
-        router = QueryRouterAgent(self.deps)
-
-        test_cases = [
-            ("Create a BWA tool", "custom_tool"),
-            ("Why did my job fail?", "error_analysis"),
-            ("What tools can I use for RNA-seq?", "tool_recommendation"),
-            ("How do I align sequences?", "tool_recommendation"),
-            ("Find a tutorial on RNA-seq", "gtn_training"),
-        ]
-
-        for query, expected_agent in test_cases:
-            decision = await router.route_query(query)
-            assert (
-                decision.primary_agent == expected_agent
-            ), f"Query '{query}' should route to {expected_agent}, got {decision.primary_agent}"
-            # Just verify we got a reasoning string, don't check exact phrasing (LLM variability)
-            assert len(decision.reasoning) > 0, f"Query '{query}' should have reasoning"
-
-    @pytest.mark.asyncio
-    async def test_router_orchestration_detection_conservative(self):
-        """Test that router orchestration detection is appropriately conservative."""
-        router = QueryRouterAgent(self.deps)
-
-        # Test cases that should NOT trigger orchestration (conservative behavior)
-        non_orchestration_cases = [
-            "What tools can I use for RNA-seq analysis?",  # Single domain: tool_recommendation
-            "My BWA job failed with error code 1",  # Single domain: error_analysis
-            "How do I create a custom Galaxy tool?",  # Single domain: custom_tool
-            "Find me tutorials about RNA-seq",  # Single domain: gtn_training
-            "I need help with my dataset format",  # Single domain: dataset_analyzer
-            "Tell me about alignment tools",  # Single domain: tool_recommendation
-            "My tool crashed",  # Single domain: error_analysis
-            "Show me Galaxy tutorials",  # Single domain: gtn_training
-        ]
-
-        for query in non_orchestration_cases:
-            decision = await router.route_query(query)
-            assert (
-                decision.primary_agent != "orchestrator"
-            ), f"Query '{query}' should NOT trigger orchestration (got {decision.primary_agent})"
-
-    @pytest.mark.asyncio
-    async def test_router_orchestration_detection_explicit_triggers(self):
-        """Test orchestration detection with explicit multi-part requests."""
-        router = QueryRouterAgent(self.deps)
-
-        # Test cases with explicit orchestration language that SHOULD trigger orchestration
-        # These require both multiple domains AND explicit conjunction language
-        orchestration_cases = [
-            # Explicit conjunctions with multiple domains
-            "My RNA-seq tool failed and also help me find alternative tools and show me tutorials",
-            "Fix this alignment error and then recommend better tools and provide training materials",
-            "I need a complete workflow for variant calling plus also help me troubleshoot any errors",
-            # Comprehensive requests spanning multiple domains
-            "Provide a full solution for RNA-seq analysis including error handling and tutorial recommendations",
-            "Walk me through the entire process from tool selection to error troubleshooting to learning resources",
-            # Problem + solution + learning patterns
-            "Help me fix this GATK error and teach me how to prevent it in the future",
-            "Solve this memory issue and learn about best practices for large datasets",
-        ]
-
-        # Note: These might not trigger orchestration if the conservative thresholds are very high
-        # The test verifies the behavior is consistent with the current conservative tuning
-        for query in orchestration_cases:
-            decision = await router.route_query(query)
-            # Log the decision for debugging
-            print(f"Query: '{query}' -> Agent: {decision.primary_agent}, Reasoning: {decision.reasoning}")
-
-    @pytest.mark.asyncio
-    async def test_router_orchestration_scoring_logic(self):
-        """Test specific orchestration scoring patterns."""
-        from galaxy.agents.router import QueryRouterAgent
-
-        router = QueryRouterAgent(self.deps)
-
-        # Test high-confidence multi-domain queries
-        multi_domain_cases = [
-            {
-                "query": "My FastQC tool failed with memory error and I need alternative quality control tools and tutorials on quality assessment",
-                "expected_domains": ["error_analysis", "tool_recommendation", "gtn_training"],
-                "should_orchestrate": True,  # 3+ high-confidence domains
-                "reason": "3+ high-confidence domains should trigger orchestration",
-            },
-            {
-                "query": "Create a BWA tool and also provide training materials",
-                "expected_domains": ["custom_tool", "gtn_training"],
-                "should_orchestrate": False,  # Only 2 domains, needs explicit language for orchestration
-                "reason": "2 domains without very explicit orchestration language should not trigger",
-            },
-        ]
-
-        for case in multi_domain_cases:
-            decision = await router.route_query(str(case["query"]))
-
-            is_orchestrated = decision.primary_agent == "orchestrator"
-
-            # Log detailed information for debugging
-            print(f"\nQuery: {case['query']}")
-            print(f"Expected orchestration: {case['should_orchestrate']}")
-            print(f"Actual agent: {decision.primary_agent}")
-            print(f"Reasoning: {decision.reasoning}")
-
-            # The assertion depends on current conservative tuning
-            # This test documents the current behavior rather than enforcing it
-            if case["should_orchestrate"]:
-                # Might still not orchestrate due to conservative thresholds
-                print(
-                    f"Expected orchestration but got {decision.primary_agent} - this may be due to conservative tuning"
-                )
-            else:
-                assert not is_orchestrated, case["reason"]
-
-    @pytest.mark.asyncio
-    async def test_router_orchestration_explicit_indicators(self):
-        """Test specific orchestration indicator patterns."""
-        router = QueryRouterAgent(self.deps)
-
-        # Test explicit orchestration language patterns
-        explicit_patterns = [
-            "and also help me",
-            "and then show me",
-            "plus also provide",
-            "as well as give me",
-            "complete workflow for",
-            "full solution including",
-            "entire process from",
-            "comprehensive help with",
-            "step by step workflow",
-            "start to finish guide",
-            "beginning to end process",
-        ]
-
-        # Create test queries that include these patterns
-        for pattern in explicit_patterns:
-            # Create a query that has the pattern but might not have enough domain signals
-            test_query = f"I need help with RNA-seq analysis {pattern} tutorials"
-
-            decision = await router.route_query(test_query)
-
-            # Log the results - these might not trigger orchestration due to conservative scoring
-            print(f"Pattern '{pattern}' in query: {decision.primary_agent}")
-
-            # The presence of explicit language should at least contribute to scoring
-            # But conservative thresholds might still prevent orchestration
-
-    @pytest.mark.asyncio
-    async def test_custom_tool_agent_structured_output(self):
-        """Test custom tool agent with structured output support."""
-        # Test with a model that supports structured output (gpt-4o)
-        self.mock_config.ai_model = "gpt-4o"
-        agent = CustomToolAgent(self.deps)
-
-        # Mock the agent run to return a UserToolSource
-        with mock.patch.object(agent.agent, "run") as mock_run:
-            mock_tool = UserToolSource(
-                **{
-                    "class": "GalaxyUserTool",
-                    "id": "test-tool",
-                    "name": "Test Tool",
-                    "version": "1.0.0",
-                    "description": "A test tool",
-                    "container": "ubuntu:latest",
-                    "shell_command": "echo test",
-                    "inputs": [],
-                    "outputs": [],
-                }
-            )
-
-            mock_result = mock.Mock()
-            mock_result.output = mock_tool
-            mock_result.data = mock_tool
-            mock_run.return_value = mock_result
-
-            response = await agent.process("Create a test tool")
-
-            assert response.confidence.value in ["high", "medium"]
-            assert response.metadata["tool_id"] == "test-tool"
-            assert response.metadata["method"] == "structured"
-
-    @pytest.mark.asyncio
-    async def test_custom_tool_agent_requires_structured_output(self):
-        """Test custom tool agent returns helpful error when model doesn't support structured output."""
-        # Test with DeepSeek which doesn't support structured output
-        self.mock_config.ai_model = "deepseek-r1"
-        agent = CustomToolAgent(self.deps)
-
-        response = await agent.process("Create a BWA-MEM tool")
-
-        # Should return capability error, not attempt fallback
-        assert response.metadata.get("error") == "model_capability"
-        assert response.metadata.get("requires") == "structured_output"
-        assert "structured output" in response.content.lower()
-        assert response.confidence.value == "low"
-
-    def test_agent_registry(self):
-        """Test that all required agents are registered."""
-        required_agents = [
-            "router",
-            "custom_tool",
-            "error_analysis",
-            "tool_recommendation",
-            "gtn_training",
-        ]
-
-        for agent_type in required_agents:
-            assert agent_registry.is_registered(agent_type), f"Agent {agent_type} should be registered"
-            # Verify we can get agent info
-            info = agent_registry.get_agent_info(agent_type)
-            assert info["agent_type"] == agent_type
-            assert "class_name" in info
-
-    def test_gtn_database_connectivity(self):
-        """Test that GTN database is accessible."""
-        from galaxy.agents.gtn import GTNSearchDB
-
-        try:
-            db = GTNSearchDB()
-            metadata = db.get_metadata()
-            assert metadata is not None
-            assert "version" in metadata or "tutorial_count" in metadata
-
-            # Test basic search
-            results = db.search("RNA-seq", limit=1)
-            assert isinstance(results, list)
-        except Exception as e:
-            # GTN database might not be available in test environment
-            pytest.skip(f"GTN database not available: {e}")
-
-    @pytest.mark.asyncio
-    async def test_gtn_agent_basic(self):
-        """Test GTN training agent basic functionality."""
-        from galaxy.agents.gtn_training import GTNTrainingAgent
-
-        agent = GTNTrainingAgent(self.deps)
-
-        # Test with a basic query
-        response = await agent.process("How do I analyze RNA-seq data?")
-
-        assert response.content is not None
-        assert response.agent_type == "gtn_training"
-        # Should recommend GTN tutorials
-        assert "tutorial" in response.content.lower() or "training" in response.content.lower()
-
-    @pytest.mark.skip(reason="TestModel API changed in pydantic-ai, needs update for new version")
-    @pytest.mark.asyncio
-    async def test_router_with_test_model(self):
-        """Test router using pydantic-ai TestModel for deterministic output."""
-        # TODO: Update this test for newer pydantic-ai TestModel API
-        # The TestModel API changed and no longer has set_result()
-        with patch("galaxy.agents.router.QueryRouterAgent._create_agent") as mock_create:
-            from pydantic_ai import Agent
-
-            # Create TestModel with predictable output
-            test_model = TestModel()
-            # This API no longer exists in newer pydantic-ai versions
-            # test_model.set_result({...})
-
-            test_agent = Agent(
-                "test-router",
-                model=test_model,
-                result_type=Dict[str, Any],
-            )
-            mock_create.return_value = test_agent
-
-            router = QueryRouterAgent(self.deps)
-            decision = await router.route_query("Create a BWA tool")
-
-            assert decision.primary_agent == "custom_tool"
-            # confidence is a Literal["low", "medium", "high"], just verify it's set
-            assert decision.confidence is not None
-
-    @pytest.mark.asyncio
-    async def test_workflow_orchestrator_agent_mocked(self):
-        """Test WorkflowOrchestratorAgent with mocked responses."""
-        from galaxy.agents.orchestrator import (
-            AgentPlan,
-            WorkflowOrchestratorAgent,
-        )
-
-        agent = WorkflowOrchestratorAgent(self.deps)
-
-        # Test 1: Query that should NOT trigger orchestration (single agent)
-        with patch.object(agent, "_get_agent_plan") as mock_get_plan:
-            # Mock a plan that indicates single agent is sufficient
-            mock_get_plan.return_value = AgentPlan(
-                agents=["tool_recommendation"],
-                sequential=False,
-                reasoning="Single tool recommendation needed",
-            )
-
-            # Mock the actual agent call to avoid running it
-            with patch("galaxy.agents.agent_registry.get_agent") as mock_get_agent:
-                mock_tool_agent = AsyncMock()
-                mock_tool_agent.process.return_value = MagicMock(
-                    content="Use BWA-MEM for alignment.",
-                    agent_type="tool_recommendation",
-                )
-                mock_get_agent.return_value = mock_tool_agent
-
-                response = await agent.process("I need tools for RNA-seq analysis")
-
-                # Should not orchestrate, just return single agent response
-                assert response.agent_type == "orchestrator"
-                assert response.metadata.get("agents_used") == ["tool_recommendation"]
-                assert "BWA-MEM" in response.content
-
-    @pytest.mark.asyncio
-    async def test_workflow_orchestrator_sequential_execution(self):
-        """Test orchestrator sequential workflow execution."""
-        from galaxy.agents.orchestrator import (
-            AgentPlan,
-            WorkflowOrchestratorAgent,
-        )
-
-        agent = WorkflowOrchestratorAgent(self.deps)
-
-        # Mock a complex plan requiring sequential orchestration
-        complex_plan = AgentPlan(
-            agents=["error_analysis", "tool_recommendation", "gtn_training"],
-            sequential=True,
-            reasoning="Multi-step workflow: error diagnosis -> tool alternatives -> learning resources",
-        )
-
-        # Mock each agent call in the sequential workflow
-        with (
-            patch.object(agent, "_get_agent_plan") as mock_get_plan,
-            patch("galaxy.agents.agent_registry.get_agent") as mock_get_agent,
-        ):
-            mock_get_plan.return_value = complex_plan
-
-            # Mock individual agent responses
-            mock_error_agent = AsyncMock()
-            mock_error_agent.process.return_value = MagicMock(
-                content="Tool failed due to memory issues", agent_type="error_analysis"
-            )
-
-            mock_tool_agent = AsyncMock()
-            mock_tool_agent.process.return_value = MagicMock(
-                content="Alternative tools: HISAT2, STAR", agent_type="tool_recommendation"
-            )
-
-            mock_training_agent = AsyncMock()
-            mock_training_agent.process.return_value = MagicMock(
-                content="Training available: RNA-seq tutorial", agent_type="gtn_training"
-            )
-
-            # Configure mock to return different agents
-            def get_agent_side_effect(agent_type, deps):
-                if agent_type == "error_analysis":
-                    return mock_error_agent
-                elif agent_type == "tool_recommendation":
-                    return mock_tool_agent
-                elif agent_type == "gtn_training":
-                    return mock_training_agent
-                else:
-                    raise ValueError(f"Unexpected agent type: {agent_type}")
-
-            mock_get_agent.side_effect = get_agent_side_effect
-
-            response = await agent.process(
-                "My RNA-seq tool failed with memory error and I need alternatives and tutorials"
-            )
-
-            # Verify orchestration occurred
-            assert response.agent_type == "orchestrator"
-            assert response.metadata.get("execution_type") == "sequential"
-            assert "memory issues" in response.content
-            assert "Alternative tools" in response.content
-            assert "Training available" in response.content
-
-            # Verify agents were called in sequence
-            assert mock_error_agent.process.called
-            assert mock_tool_agent.process.called
-            assert mock_training_agent.process.called
-
-    @pytest.mark.asyncio
-    async def test_workflow_orchestrator_parallel_execution(self):
-        """Test orchestrator parallel workflow execution."""
-        from galaxy.agents.orchestrator import (
-            AgentPlan,
-            WorkflowOrchestratorAgent,
-        )
-
-        agent = WorkflowOrchestratorAgent(self.deps)
-
-        # Mock parallel plan
-        parallel_plan = AgentPlan(
-            agents=["tool_recommendation", "gtn_training"],
-            sequential=False,
-            reasoning="Independent tasks can run in parallel",
-        )
-
-        with (
-            patch.object(agent, "_get_agent_plan") as mock_get_plan,
-            patch("galaxy.agents.agent_registry.get_agent") as mock_get_agent,
-        ):
-            mock_get_plan.return_value = parallel_plan
-
-            # Mock agent responses
-            mock_tool_agent = AsyncMock()
-            mock_tool_agent.process.return_value = MagicMock(
-                content="Recommended tools: BWA, Bowtie2", agent_type="tool_recommendation"
-            )
-
-            mock_training_agent = AsyncMock()
-            mock_training_agent.process.return_value = MagicMock(
-                content="Available tutorials: Alignment workflow", agent_type="gtn_training"
-            )
-
-            def get_agent_side_effect(agent_type, deps):
-                if agent_type == "tool_recommendation":
-                    return mock_tool_agent
-                elif agent_type == "gtn_training":
-                    return mock_training_agent
-                else:
-                    raise ValueError(f"Unexpected agent type: {agent_type}")
-
-            mock_get_agent.side_effect = get_agent_side_effect
-
-            response = await agent.process("I need alignment tools and training materials")
-
-            # Verify parallel execution
-            assert response.agent_type == "orchestrator"
-            assert response.metadata.get("execution_type") == "parallel"
-            assert "Recommended tools" in response.content
-            assert "Available tutorials" in response.content
-
-    @pytest.mark.asyncio
-    async def test_workflow_orchestrator_fallback_behavior(self):
-        """Test orchestrator fallback when planning fails."""
-        from galaxy.agents.orchestrator import WorkflowOrchestratorAgent
-
-        agent = WorkflowOrchestratorAgent(self.deps)
-
-        # Mock planning failure
-        with patch.object(agent, "_get_agent_plan") as mock_get_plan:
-            mock_get_plan.side_effect = Exception("LLM service unavailable")
-
-            response = await agent.process("Complex query that should trigger fallback")
-
-            # Should fall back gracefully
-            assert response.agent_type == "orchestrator"
-            assert "having trouble" in response.content
-
-
-# ============================================================================
-# LIVE LLM TEST SUITE - Requires configured LLM
-# ============================================================================
-
-
-@pytestmark_live_llm
-class TestAgentsApiLiveLLM(ApiTestCase):
-    """Test Galaxy AI agents API with real LLM.
-
-    These tests require a configured LLM and will be skipped unless
-    GALAXY_TEST_ENABLE_LIVE_LLM=1 is set.
-    """
-
-    def setUp(self):
-        super().setUp()
-        self.dataset_populator = DatasetPopulator(self.galaxy_interactor)
-        self.workflow_populator = WorkflowPopulator(self.galaxy_interactor)
-
-    def test_query_agent_auto_routing_live(self):
-        """Test automatic agent routing with live LLM."""
-        response = self._post(
-            "ai/agents/query",
-            data={
-                "query": "Create a BWA-MEM tool for paired-end reads",
-                "agent_type": "auto",
-            },
-            json=True,
-        )
-        self._assert_status_code_is_ok(response)
-        data = response.json()
-        assert "response" in data
-        assert "agent_type" in data["response"]
-        # Router should route this to custom_tool
-        assert data.get("routing_info", {}).get("selected_agent") == "custom_tool"
-
-    def test_query_custom_tool_agent_live(self):
-        """Test custom tool agent with live LLM."""
-        response = self._post(
-            "ai/agents/query",
-            data={
-                "query": "Create a simple tool that counts lines in a file",
-                "agent_type": "custom_tool",
-            },
-            json=True,
-        )
-        self._assert_status_code_is_ok(response)
-        data = response.json()
-        assert "response" in data
-        agent_response = data["response"]
-        assert "content" in agent_response
-        assert "metadata" in agent_response
-        assert "tool_id" in agent_response["metadata"]
-        assert "tool_yaml" in agent_response["metadata"]
-        # Check that it created something sensible
-        tool_yaml = agent_response["metadata"]["tool_yaml"]
-        assert "command" in tool_yaml or "shell_command" in tool_yaml
-
-    def test_error_analysis_endpoint_live(self):
-        """Test the dedicated error-analysis endpoint."""
-        response = self._post(
-            "ai/agents/error-analysis",
-            data={
-                "query": "My BWA job failed with exit code 137 and stderr shows 'Killed'",
-                "error_details": {"tool_id": "bwa_mem", "exit_code": 137},
-            },
-            json=True,
-        )
-        self._assert_status_code_is_ok(response)
-        data = response.json()
-        assert "content" in data
-        assert "confidence" in data
-        # Should mention memory or OOM since exit code 137 is SIGKILL
-        content = data["content"].lower()
-        assert any(word in content for word in ["memory", "kill", "resource", "oom"])
-
-    def test_tool_recommendation_endpoint_live(self):
-        """Test the dedicated tool-recommendation endpoint."""
-        response = self._post(
-            "ai/agents/tool-recommendation",
-            data={
-                "query": "I have paired-end FASTQ files and want to align them to a reference genome",
-                "input_format": "fastq",
-            },
-            json=True,
-        )
-        self._assert_status_code_is_ok(response)
-        data = response.json()
-        assert "content" in data
-        assert "confidence" in data
-        # Should mention alignment tools
-        content = data["content"].lower()
-        assert any(word in content for word in ["bwa", "bowtie", "align", "map"])
-
-    def test_custom_tool_endpoint_live(self):
-        """Test the dedicated custom-tool endpoint."""
-        response = self._post(
-            "ai/agents/custom-tool",
-            data={
-                "query": "Create a tool that counts sequences in a FASTA file",
-            },
-            json=True,
-        )
-        self._assert_status_code_is_ok(response)
-        data = response.json()
-        assert "content" in data
-        assert "metadata" in data
-        # Should generate tool YAML
-        metadata = data["metadata"]
-        assert "tool_yaml" in metadata or "tool_id" in metadata
-
-    def test_chat_endpoint_live(self):
-        """Test the chat endpoint with auto routing."""
-        response = self._post(
-            "chat?query=What%20tools%20are%20available%20for%20RNA-seq%3F&agent_type=auto",
-            data={},
-            json=True,
-        )
-        self._assert_status_code_is_ok(response)
-        data = response.json()
-        assert "response" in data
-        # Should return some content about RNA-seq
-        assert len(data["response"]) > 0
-
-    def test_chat_history_endpoint(self):
-        """Test the chat history endpoint."""
-        response = self._get("chat/history?limit=5")
-        self._assert_status_code_is_ok(response)
-        data = response.json()
-        # Should return a list (may be empty)
-        assert isinstance(data, list)
-
-
-@pytestmark_live_llm
-class TestAgentUnitLiveLLM:
-    """Unit tests with real LLM connections."""
-
-    def setup_method(self):
-        """Set up real dependencies for live LLM testing."""
-        self.mock_config = mock.Mock()
-        self.mock_config.ai_api_key = os.environ.get("GALAXY_AI_API_KEY", "test-key")
-        self.mock_config.ai_model = os.environ.get("GALAXY_AI_MODEL", "llama-4-scout")
-        self.mock_config.ai_api_base_url = os.environ.get("GALAXY_AI_API_BASE_URL", "http://localhost:4000/v1/")
-
-        self.mock_user = mock.Mock()
-        self.mock_user.id = 1
-        self.mock_user.username = "test_user"
-
-        self.mock_trans = mock.Mock()
-        self.mock_trans.app.config = self.mock_config
-        self.mock_trans.user = self.mock_user
-
-        self.deps = GalaxyAgentDependencies(
-            trans=self.mock_trans,
-            user=self.mock_user,
-            config=self.mock_config,
-            job_manager=None,
-        )
-
-    @pytest.mark.asyncio
-    async def test_router_agent_routing_decisions_live(self):
-        """Test router with real LLM."""
-        router = QueryRouterAgent(self.deps)
-
-        test_cases = [
-            ("Create a BWA tool", "custom_tool"),
-            ("Why did my job fail?", "error_analysis"),
-            ("What tools can I use for RNA-seq?", "tool_recommendation"),
-        ]
-
-        for query, expected_agent in test_cases:
-            decision = await router.route_query(query)
-            assert (
-                decision.primary_agent == expected_agent
-            ), f"Query '{query}' should route to {expected_agent}, got {decision.primary_agent}"
-
-    @pytest.mark.asyncio
-    async def test_custom_tool_agent_with_scout(self):
-        """Test custom tool agent with Scout model."""
-        self.mock_config.ai_model = "llama-4-scout"
-        agent = CustomToolAgent(self.deps)
-
-        response = await agent.process("Create a simple echo tool")
-
-        assert response.confidence in ["high", "medium"]
-        assert "tool_id" in response.metadata
-        assert "tool_yaml" in response.metadata
-        assert response.metadata["method"] == "structured"
-
-    @pytest.mark.asyncio
-    async def test_custom_tool_agent_with_deepseek(self):
-        """Test custom tool agent with DeepSeek model."""
-        self.mock_config.ai_model = "deepseek-r1"
-        agent = CustomToolAgent(self.deps)
-
-        response = await agent.process("Create a simple echo tool")
-
-        # DeepSeek should use fallback
-        assert response.metadata["method"] == "simple_template"
-        assert "tool_id" in response.metadata
-        assert "tool_yaml" in response.metadata
-
-
-@pytestmark_live_llm
-class TestAgentConsistencyLiveLLM:
-    """Test agents with a consistent set of questions."""
-
-    TEST_QUERIES = [
-        # Tool creation queries
-        ("Create a simple line counting tool", "custom_tool"),
-        ("Build a Galaxy tool that runs samtools sort", "custom_tool"),
-        ("I need a wrapper for BWA-MEM", "custom_tool"),
-        # Error analysis queries
-        ("Why did my job fail with exit code 127?", "error_analysis"),
-        ("Help me debug this memory error", "error_analysis"),
-        ("What does 'command not found' mean?", "error_analysis"),
-        # Tool recommendation queries
-        ("What tools can I use for RNA-seq analysis?", "tool_recommendation"),
-        ("How do I convert BAM to FASTQ?", "tool_recommendation"),
-        ("Which aligner should I use for short reads?", "tool_recommendation"),
-        # GTN training queries
-        ("Find me a tutorial on variant calling", "gtn_training"),
-        ("How do I learn Galaxy basics?", "gtn_training"),
-        ("Training materials for proteomics", "gtn_training"),
-        # Meta queries (should get direct response)
-        ("Hello", None),
-        ("Thank you", None),
-        ("What can you do?", None),
-    ]
-
-    @pytest.fixture
-    def live_deps(self):
-        """Create dependencies for live LLM testing."""
-        mock_config = mock.Mock()
-        mock_config.ai_api_key = os.environ.get("GALAXY_AI_API_KEY", "test-key")
-        mock_config.ai_model = os.environ.get("GALAXY_AI_MODEL", "llama-4-scout")
-        mock_config.ai_api_base_url = os.environ.get("GALAXY_AI_API_BASE_URL", "http://localhost:4000/v1/")
-
-        mock_user = mock.Mock()
-        mock_user.id = 1
-        mock_user.username = "test_user"
-
-        mock_trans = mock.Mock()
-        mock_trans.app.config = mock_config
-        mock_trans.user = mock_user
-
-        return GalaxyAgentDependencies(
-            trans=mock_trans,
-            user=mock_user,
-            config=mock_config,
-            job_manager=None,
-        )
-
-    @pytest.mark.asyncio
-    async def test_routing_consistency_live(self, live_deps):
-        """Test that routing is consistent for known query types with live LLM."""
-        router = QueryRouterAgent(live_deps)
-
-        for query, expected_agent in self.TEST_QUERIES:
-            decision = await router.route_query(query)
-
-            if expected_agent is None:
-                # Should provide direct response
-                assert decision.direct_response is not None, f"Query '{query}' should get direct response"
-            else:
-                # Should route to expected agent
-                assert (
-                    decision.primary_agent == expected_agent
-                ), f"Query '{query}' should route to {expected_agent}, got {decision.primary_agent}"
-
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize("query,expected_agent", TEST_QUERIES)
-    async def test_individual_query_routing_live(self, live_deps, query, expected_agent):
-        """Test each query individually with live LLM."""
-        router = QueryRouterAgent(live_deps)
-        decision = await router.route_query(query)
-
-        if expected_agent is None:
-            assert decision.direct_response is not None
-        else:
-            assert decision.primary_agent == expected_agent
-
-
-# Command-line test runner
-if __name__ == "__main__":
-    import sys
-
-    # Check if we should run live LLM tests
-    run_live = os.environ.get("GALAXY_TEST_ENABLE_LIVE_LLM")
-
-    print("Running Galaxy Agent Tests")
-    print("=" * 60)
-
-    if run_live:
-        print("Mode: LIVE LLM (using real LLM connections)")
-        print(f"Model: {os.environ.get('GALAXY_AI_MODEL', 'default')}")
-        print(f"Endpoint: {os.environ.get('GALAXY_AI_API_BASE_URL', 'default')}")
-    else:
-        print("Mode: MOCKED (deterministic, no LLM required)")
-        print("To run live tests: export GALAXY_TEST_ENABLE_LIVE_LLM=1")
-
-    print("=" * 60)
-
-    # Run appropriate test suite
-    if run_live:
-        # Run live LLM tests
-        pytest_args = [
-            __file__,
-            "-v",
-            "-k",
-            "LiveLLM",
-            "--tb=short",
-        ]
-    else:
-        # Run mocked tests
-        pytest_args = [
-            __file__,
-            "-v",
-            "-k",
-            "Mocked",
-            "--tb=short",
-        ]
-
-    print(f"\nRunning: pytest {' '.join(pytest_args)}\n")
-    sys.exit(pytest.main(pytest_args))
